Update patch set 1
No score.

I've talked to hackbod and now understand the backstory here better. 
I'm fine with the change, though I'd like to see a comment in the code
about why we're using two different power-state broadcasts -- yes, there
are a couple of very good reasons, but documentation is good!  

Patch-set: 1