Update patch set 1
No score.

I made a similar comment in 1742, but perhaps it is best addressed here:

Would it be better to suggest that FreeBSD users, and any other system
with non-Linux file locations, create a symlink to bash at /bin/bash?
Perhaps the script could check for the existence of /bin/bash and inform
the user if it does not exist?

Patch-set: 1