Slim down chooseproduct (used by choosecombo).
People keep getting confused by the default products
in this list, which are almost never what they want,
because this list roughly corresponds to the DEVICES
and not the PRODUCTS.

Now that lunch works again, I'm not too worried about
choosecombo except for the couple of cases lunch can't
handle, like the simulator.  Therefore, this change
just removes the list of suggested products, and you
have to know which one you're trying to build (which
seems like a resonable prerequisite).