Implements the basic structure for providing localized documentation; the site tabs are now in a separate file to ease the translation process, the language selection menu has been added to the header, scripts have been added to dynamically switch the language of the site tabs and dev guide side navigation and save the language preference in a browser cookie, and some of the functions in macro.cs have moved into customization.cs.
patch 2:
revised the logic for changing tab and nav languages so that,
in the event that a SPAN node exists for the newly selected language
but the translation has not actually been put there, it will not be selected
and the previously selected language will remain for that node.

patch 3:
revised the logic introduced in patch 2, which, when loading a new page
would actually reveal all languages if the selected language had no
translation. Now, we must loop through each occurance and check whether
a translation is available. If not, then we hide it and show English.

patch 4:
small error fix