Added support for TARGET_ARCH_VERSION=armv5te-vfp.
This is currently a copy & paste of the armv5te parameters.  I don't
know if there's a better way to encode this, or to what extent we will
need to specialize it vs. armv5te in a future build (e.g. to enable fp
instruction use in code gen).

Currently the only effect is to select the VFP-enabled mterp sources in
Dalvik.