CTS: Add missing integration of 147397 to fix VM tests
For host side test packages, the name of the jar file containing the tests is
declared in the packages Android.mk.  The name of the make file needs to be
passed to the description generator so that it can pick up this information.
The makefile was not passing on the parameter containing the make file name.