Fix ARMv7 -user builds
Allow the dex preoptimization pass for ARMv7 system images to use the
ARMv7-specific kernel image. This is needed to fix the -user builds that
absolutely depend on running the dex preopt in the emulator.

This will hopefully be removed when we'll be able to perform the dex
preopt on the host instead.