docs: add checkbox to toggle the api level filter on and off; this includes some function renaming for the related scripts.
the checkbox will be unchecked by default, so any users
that currently have an api level filter selected will now
see all apis until they click the checkbox. the old api level
cookie is still saved the same, but the selected value will only
be applied if the checkbox is selected, which has its own cookie
for on or off.