Add a host shell function to generate system_server stack traces
For debugging non-ANR weird states in the system server, it is incredibly handy
to be able to force a complete thread-stack dump of that process.  Here's a
shell function that does so (creating the /data/anr/traces.txt file with the
usual permissions if it does not yet exist).

Depends on adb having been configured to run as root.

Change-Id: I01d9d733ef08ffefb00cb61968df36ecf62087f0