Fix a Windows-only discrepency in the Windows SDK build.
For some reason, there's a slight behavior difference between
my Windowss machine and the Windows build servers, that is on
mine the include of framework/base results in duplicate build
rules being defined. My machine used to work like the Windows
build server and the behavior mysteriously changed a while
ago. This change should make it all work the same on all boxes.

Change-Id: I45e02e9eadfc586f86caa6a2690deda9648e600d