Remove special-case where CLASS=apps ignored TAGS=user
Also add a warning when there's TAG=user on a CLASS=apps module,
as this is not what people normally intend (this should be handled
in core.mk instead).

This used to be a source of confusion, with people using TAGS=user on
APPS and non-APPS modules, and being surprised to see some of those
included but not others.

Change-Id: I56b5618a556abdc6b841bf7a943d77c20fe9a4c5