Manual merge of change 50994 from master (DO NOT MERGE)
Merging from master because I'm bringing my maps API work into
gingerbread.

Fix and enable JarJar in Java builds.

java.mk contained a target to enable JarJar in Java builds, but it was
being bypassed.  This change enabled JarJar as part of the build
process, and fixes a slight bug with the way it was set up.

Change-Id: I814e64eb30b41bf249a63d683c03040e2658fe0d