Fixed a bunch of stuff in the prelink mappings file:
* Added mappings for most of the libraries that didn't already
have them (as observed from my build).

* Re-canonicalized all the offsets in the mappings file,
per the official policy of keeping things megabyte-aligned.

* Added comments on each line indicating what size the library
actually appeared to be in my build (to help with future rejiggerings).

* Added a bit of documentation about the whole deal.

Even with this change, I know there are other libraries that will
still need to be included, but I think what I have here is a good
step. Also, this change only reserves room for the new libraries, but
subsequent changes (in other subprojects) are required to actually
start prelinking the new libraries. Those changes will be forthcoming
as well.

Bug: 2651381
Change-Id: I7a06382fc2b8a14bbd3261de58df3f8a31e4108e