Update patch set 4
Patch Set 4: Fails

Unfortunately this change can't be merged into Google's master branch (which is where it needs to go in order to be part of official platform releases and not get eventually deleted), because the config.mk and combo/ mechanism is different there.

Once we open-source a source tree with the new logic, we should revisit this issue.

Patch-set: 4
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1