Instrumentation links against the jar file for its app, so use that at the dependency.
I don't remember exactly, but I think the dependencies for apps
were broken, and you couldn't reliably cause the java to be rebuilt
by depending just on the jar file-- you needed to depend on the
apk.  That isn't the case anymore, and relying on the apk means
that something built in out/target/common was depending on something
that's deleted by installclean.

With this change (and others), on my mac, the build after
installclean for passion-eng takes 2 minutes instead of 11.

Change-Id: Ib490f88292d591254013e07958961fc50ec2e3dc