Fixed problem with keymap files copied to target
Moved Android.mk and keymap files except AVRCP.kl from sdk/emulator/keymaps
to build/target/board/generic. This causes the keymap files to copied into
the system image of generic and sdk products, but not into the system image
of HW products. Originally the qwerty.kl was copied to target no matter
if the target had a qwerty keyboard or not, which is the case for e.g.
the Nexus One.

Change-Id: I7cd80497632fb3b9e7ce2f3a7888d4924a355f52