Use the new sdk/build/tools.atree.
As the sdk.git project is branched differently from the rest
of the platform, it's important to be able to control what
goes in the tools folder of the SDK on a per-branch basis.

Right now development/build/sdk.atree control what goes in
the tools folder (coming almost only from sdk.git) which causes
problems when the development.git and sdk.git branches are not
in sync.

Change-Id: Ia0b0ea37f8a4904b3a7f7805a552eeb499a2c0ee