Improve setup of base products.
Split telephony into its own product, so that we can easily
add telephony support to any configuration without pulling
anything else. Use it from generic.

Split the "full" family between a (no-telephony) base,
a telephony version of the same (both of which can be used
as a starting point for on-device builds), and the true classic
full which is an emulator build.

Change-Id: I616d34f2a29415b8ec997173f8e893c312ade5c2