Added generic hook for loading vendor specific tasks.
With this change it is possible for vendors to create rules that depend on
targets defined in Android.mk files.

This makes it easier to decouple functionality from the core buildsystem.
Very much as is done within the build git for tasks like creating SDK addon,
Generating eclipse .classpath file etc can then be done for vendors as well.

You can add targets that could perform poststeps on produced modules.
E.g. Do static verification or documentation on modules of a certain type.
You can e.g. create a target that would resign and upload all apks to a
binary repository.

Change-Id: Ic0fa7b60b919ede3c59fa2885d1fd4be1e928c04