core/binary.mk: Add missing Objective-C objects to the build. DO NOT MERGE
NOTE: This is a back-port from AOSP into the internal gingerbread branch.

For some reason, all_objects was not listing objects generated from
objective-C sources. We need these for the emulator (which builds its
own version of SDL which uses Quartz on OS X).

+ Fix a typo in transform-host-m-to-o definition.

This also explains why https://review.source.android.com/#change,21074
had to be reverted, since it gets rid of the prebuilt SDL libraries when
building the emulator.

Change-Id: I03e179f3216ed95908bf8e8011d34109accff1b9
Orig-Change-Id: I173811cf11cdb5b045073aade59364236145bc77