Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

(1 inline comment)

The files here are purposefully a clone of the generic/* files. They build for the same target (goldfish emulator)... but this version builds for ATOM.

It's useful IMO to complete sync these files with the generic/* versions.

Patch-set: 5
Label: Code-Review=+1