Support target-specific sdk/build/tools.atree extension.
The purpose of this patch is to add the ability to specify a
target-specific extension to sdk/build/tools.atree, under the
name sdk/build/tools.$(TARGET_ARCH).atree

This is needed to move x86-specific changes out of tools.atree
in order to fix the build of internal Android branches that
don't include other x86-related changes to the build system.

Another patch, following this patch, will fix the build itself
by moving the x86-stuff to sdk/build/tools.x86.atree.

NOTE: The root cause of the problem is that tools.atree
probably shouldn't be in the public AOSP repository, but
under development/build/ instead. However, fixing this
requires more drastic changes to the build system.

Change-Id: Ie365c55527bcad38e1e5248f618e2fd5a96431a1