full_x86-eng: Ensure we copy the PC bios files.
This change ensures that building full_x86-eng will copy the
Bios files to the right location. This is necessary to launch
"emulator-x86" after building the system images.

Before this patch, one had to make the copy manually. Note that
sdk_x86-eng does not need this (the files are actually taken care
of in build/target/product/sdk.mk so will be copied even in case
of the ARM sdk build).

Change-Id: If70adead27c7997694ebc30561005d26a7358c34