Added audio effects config file.
A configuration file named audio_effects.conf is now used
to list audio effect libraries and engines to be loaded by the
effects factory. A default version of this file is installed
in /system/etc. OEMs can override this file by placing a file
with the same name under /vendor/etc.

Change-Id: Id10e0e54f53234ec52b22b6f882734edaf2648ac