Update patch set 4
Patch Set 4:

The gcc driver "should" do a number of things that are being done in this file. It's very inconvenient to have the makefiles have to "know" how to find the crt files and set lots of flags that "gcc -mandroid" really should do.

Last year, we were told that Google wanted to move away from -mandroid and from modifying the gcc driver.... yet... the question keeps coming back when we work around the lack of support in the driver.

I'm fine with the gcc driver doing much of this work. It would be very convenient. Is that the direction that Google wishes to go? Can we get the Google compiler team to make the changes?

Patch-set: 4