Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

This doesn't make sense.

String.indexOf() is almost certainly a red herring. I suspect the real problem is that you're getting the wrong string. Perhaps you're occasionally seeing docs from a superclass? Or perhaps there's a hash collision?

You should instrument the code to log the 'needle' and 'haystack' arguments when the expected string is not found. Then run it a hundred times and see what it prints when it fails. It won't be the same as the other scenarios.

Patch-set: 3