emulator: Enable dex-preoptimization
Backport from AOSP master.

This enables the Dex pre-optimization for emulator-specific builds.
The idea is to speed up the initial boot sequence of an SDK AVD by
avoiding the costly on-device preopt pass that is performed otherwise
the first time the AVD is launched.

Benchmarking shows reductions in first boot time from 1m30s to 40s
on a 2.4GHz Xeon when emulating the AOSP master ToT.

Change-Id: Ibf5ebe68aa4d363c9df05d2ffe7cb828ef99f922
Orig-Change-Id: I67a9ce9791220c202817864c6e83dc8fd9cbc027