docs: fix xss vulnerability bug 5008593
the fix deletes some code that was originally there to handle side nav
highlighting for inter-doc links. This code is no longer necessary (or at least
should not be) because side nav links should not be inter-doc links, but
should always be to their own document (no two sidenav links should go
to separate sections of the same doc).

Change-Id: If77d6fe869e6a44bae2f194c9de26222e6ad9419