Use LOCAL_REQUIRED_MODULES to add dependency of installed shared libraries
Bug: 5200343
Before this change the build system assumes that shared libraries will
be installed to $(TARGET_SHARED_LIBRARIES). That's not true if the
module uses custom LOCAL_MODULE_PATH.
With this change we defer the evaluation of the dependency to after all
Android.mks are loaded and all modules have their installed path
determined.

Change-Id: Ib6628ed44154a90e8668d31691fbbac67125be6c