Cherrypick 944311 from master. do not merge.
emulator: Enable dex-preoptimization in emulator-specific builds

This patch re-enables dex preoptimization in emulator-specific
builds. This considerably speeds up the first-boot of SDK AVDs
since the dex optimization pass is performed during the build
instead.

This was enabled in a previous patch, then reverted due to a
bug in the build system that has been fixed since.

This patch doesn't affect any device build.

Change-Id: Ie711fb4e1672993b7f7f4b8c5313a89d0b8734d6