emulator: Enable dex-preoptimization in emulator-specific builds
This patch re-enables dex preoptimization in emulator-specific
builds. This considerably speeds up the first-boot of SDK AVDs
since the dex optimization pass is performed during the build
instead.

This was enabled in a previous patch, then reverted due to a
bug in the build system that has been fixed since.

This patch doesn't affect any device build.

Change-Id: I3be8a9c7f51a86029fde30b7d37584cb40b4db69