Rename help function in envsetup.sh to be hmm
Bash has a built-in command "help" that is useful for getting
help on bash commands without wading through the bash man page
(e.g. "help test" will give you the syntax for those "if [ ]"
expressions).  Overwriting a built-in with a convenience function
is not good.

The function "hmm" was added recently that serves much the same
purpose as help.  It seems like a good idea to remove the new
"hmm" command and rename "help" to be "hmm" so we no longer
overwrite the built-in.

Change-Id: I2079b3e6a10715956ec077d31e817a939c6ee4a0
Signed-off-by: Scott Anderson <saa@android.com>