Fix rild running in Nakasi.
The module rild is telephony specific, since Nakasi has no radio
it shouldn't run rild. Changing the build system so rild is optional
for products that don't have telephony support.

Since rild is now optional, it is necessary to add it explicitly to
the sdk build as the sdk build is special.

Bug: 6463864
Change-Id: If944e051023b567807585fb91163a1e3f79c1024