Fix broken dependencies with the layers.txt files.
When a device java module was using the layers.txt files, and that device java module depenended on
a host java module, when building the host java module, the layer checker tried to run and would
fail, complaining about the host module.

The problem was that since the value wasn't set for the host module, it would be inherited from its
dependent module (a make "feature") and would cause the layer checker to run.  So this change makes
it explicitly set PRIVATE_JAVA_LAYERS_FILE for host java modules too.

Change-Id: Ifbc085295d0995a2da5d6d63d8cb379ef2631cc4