Update patch set 1
Patch Set 1:

Since -march isn't being set, won't the build simply default to the architecture of the build server... and thus do the right thing? Why is this necessary?

If the idea is to cross-build and export the host binar4ies, then you should probably be explicit with -march. If, somehow, the compiler realized that it was being built on a machine that enabled movbe... then the exported binaries could well fail on "core" processors.

Patch-set: 1