Update patch set 1
Patch Set 1:

This isn't for cross building.
Every once in a while we need to write some assembly code. The host builds are -m32 and so the fpmath defaults to x87 and SSE code/registers are disallowed. This change allows the same assembly code on both host and target.

Patch-set: 1