Don't concatenate the NOTICE file for a single module.
When the NOTICE file for a module changed, the new NOTICE was being
concatenated onto the end of the old one in non-clean builds, taking up
more and more space and presumably duplicating most of the content. Only
one set of rules for a given file can exist in make, so there cannot be
a case where more than one $(notice_file) was intentionally being
concatenated together as this would be an error.

Change-Id: I0f5fae498225eaee5cc75d854223b5ba790fcebe