Add --exact flag to `pid` tool; new `qpid` tool.
The new --exact flag for pid does an exact match on the
process name, rather than grepping anywhere in the ps
output, which helps target a specific process if its name is
a substring of another process name. (Nothing else about
pid's output, including inclusion of the ps header if it
matches, is affected.)

qpid ("quick pid" or "quiet pid") lists all processes in the
following simplified format:

	<pid> <procname>

It also helpfully strips off the header line from ps.

With an argument, qpid will search for processes or pids
matching the argument. With --exact it matches process names
exactly (as does pid, above).

Change-Id: I28a201df40281a66cb1a2918b7ee3a0b2d7b6ffd