fix sidebar autoscroll when the selected doc is not visible. bug: 9676913
This was not working for third generation list items. Switching
the position measurement to be relative to the entire page rather than
the parent container fixes it. I then added 125px to account for
the site header and set the threshold and placement of scrolled items
to be at 80% the nav height.

Change-Id: Idda1896d11069e56325ba830ed745dae9e0fc483